+++ src/migrations/m230412_233209_namespace_change_migration.php
@@ -16,13 +16,13 @@ class m230412_233209_namespace_change_migration extends Migration
     public function safeUp(): bool
     {
         echo "m230412_233209_namespace_change_migration executing...\n";
-        Craft::info("Inside of safeUp!!!!", "shavvy!");
 
         $this->update('{{%fields}}', [
-            "type" => 'lsst\dam\fields\DAMAsset',
-    ],[
+            "type" => 'lsst\dam\fields\DAMAsset'
+            ],[
             "type" => 'rosas\dam\fields\DAMAsset'
-    ]);
+            ]
+        );
 
         return true;
     }
@@ -32,7 +32,14 @@ class m230412_233209_namespace_change_migration extends Migration
      */
     public function safeDown(): bool
     {
-        echo "m230412_233209_namespace_change_migration cannot be reverted.\n";
-        return false;
+        echo "m230412_233209_namespace_change_migration is be reverted.\n";
+
+        $this->update('{{%fields}}', [
+            "type" => 'rosas\dam\fields\DAMAsset'
+        ],[
+            "type" => 'lsst\dam\fields\DAMAsset'
+        ]
+        );
+        return true;
     }
 }